+++ pyv/exception_unit.py
@@ -12,25 +12,33 @@ class ExceptionUnit(Module):
 
         self.pc_i = Input(int, [None])
         self.ecall_i = Input(bool)
+        self.mret_i = Input(bool)
 
         self.raise_exception_o = Output(bool)
         self.npc_o = Output(int)
         self.mcause_o = Output(int)
         self.mepc_o = Output(int)
+        self.trap_return_o = Output(bool)
 
         self.mepc_o << self.pc_i
 
     def process(self):
         ecall = self.ecall_i.read()
+        mret = self.mret_i.read()
         raise_ex = False
-        mtvec = 0
+        npc = 0
         mcause = 0xFFFF_FFFF
+        trap_return = False
 
         if ecall:
             raise_ex = True
-            mtvec = self.csr.read(CSR['mtvec']['addr'])
+            npc = self.csr.read(CSR['mtvec']['addr'])
             mcause = 11
+        elif mret:
+            trap_return = True
+            npc = self.csr.read(CSR['mepc']['addr'])
 
         self.raise_exception_o.write(raise_ex)
-        self.npc_o.write(mtvec)
+        self.npc_o.write(npc)
         self.mcause_o.write(mcause)
+        self.trap_return_o.write(trap_return)