+++ pyv/models/singlecycle.py
@@ -59,18 +59,21 @@ class SingleCycle(Module):
         self.bu.target_i      << self.alu_res
 
         # Connect exception logic
-        self.excep.pc_i       << self.pc
-        self.bu.raise_exception_i << self.excep.raise_exception_o
-        self.bu.mtvec_i << self.excep.npc_o
-        self.csr_unit.ex_i << self.excep.raise_exception_o
-        self.csr_unit.mepc_i << self.excep.mepc_o
-        self.csr_unit.mcause_i << self.excep.mcause_o
-        self.excep.ecall_i    << self.id_stg.ecall_o
+        self.excep.pc_i             << self.pc
+        self.excep.ecall_i          << self.id_stg.ecall_o
+        self.excep.mret_i           << self.id_stg.mret_o
+        self.bu.raise_exception_i   << self.excep.raise_exception_o
+        self.bu.mtvec_i             << self.excep.npc_o
+        self.bu.trap_return_i       << self.excep.trap_return_o
+        self.bu.mepc_i              << self.excep.npc_o
+        self.csr_unit.ex_i          << self.excep.raise_exception_o
+        self.csr_unit.mepc_i        << self.excep.mepc_o
+        self.csr_unit.mcause_i      << self.excep.mcause_o
 
         # Connect WBStage to CSR unit
-        self.csr_unit.write_en_i << self.wb_stg.csr_write_en_o
-        self.csr_unit.write_addr_i << self.wb_stg.csr_write_addr_o
-        self.csr_unit.write_val_i << self.wb_stg.csr_write_val_o
+        self.csr_unit.write_en_i    << self.wb_stg.csr_write_en_o
+        self.csr_unit.write_addr_i  << self.wb_stg.csr_write_addr_o
+        self.csr_unit.write_val_i   << self.wb_stg.csr_write_val_o
 
     def connects(self):
         val = self.IFID.read().pc