+++ test/test_singlecycle.py
@@ -121,7 +121,7 @@ class TestCSR:
         assert core.csr_unit.read(0x301) == 0x4000_0100
 
 
-class TestECALL:
+class TestExceptions:
     def test_ecall(self, sim: Simulator, core: SingleCycle):
         sim.reset()
 
@@ -139,3 +139,17 @@ class TestECALL:
         assert core.csr_unit.read(mepc_addr) == 4
         assert core.csr_unit.read(mcause_addr) == 11
         assert core.pc.read() == mtvec
+
+    def test_mret(self, sim: Simulator, core: SingleCycle):
+        sim.reset()
+
+        inst = 0x30200073
+        nop = 0x13
+        mepc_addr = 0x341
+        mepc = 16
+        core.csr_unit._dbg_set_csr(mepc_addr, mepc)
+        mem_write_word(core, 0, nop)
+        mem_write_word(core, 4, inst)
+        mem_write_word(core, 16, nop)
+        sim.run(3, False)
+        assert core.pc.read() == mepc